Fix parsing of the access_token response with ActiveRecord::JSON.
The JSON parser that is provided by ActiveRecord will not report an
exception, causing the string not to be parsed; this in turn will cause a
ripple-effect through the code (and indeed specs will fail).

With this change, the params value is guaranteed to be an Hash, and thus
work as intended.