Fix refresh token behavior.
If a refresh token isn't returned on a refresh response,
copy it over from the existing token. Google only returns
the refresh token when a user explicitly clicks to authorize,
so every subsequent time, it's not included. I think it
makes more sense to maintain all relevant attributes of
the token when it gets refreshed, and this includes the
refresh token.