Remove circular reference for OAuth2::Error
OAuth2::Error contains response and adds a link to itself from response. This is a circular reference and it can lead to various problems. E.g., attempt to dump exception to json (that most exception handlers are doing) will lead to `ActiveSupport::JSON::Encoding::CircularReferenceError: object references itself`.

This is why I propose to remove this inter-dependency.