Remove authentication form encoding
According to the OAuth 2 spec, section 2.3.1[1], the username AND the
password should be invidually encoded as `application/x-www-form-urlencoded`
(as was done in the prior commit). However, in experimentation, some
OAuth providers have rejected such aggressive encodings.

Perhaps, `URI.escape` is more appropriate, though less of a literal
interpretation of the spec. For now, I'm undoing this change as it is
orthogonal to the purpose of this stream of work (generalising
header-based authentication).

It's worth noting that the general Basic Auth specification[2], does NOT
expect encoding of username or password in this way.

[1]: https://tools.ietf.org/html/rfc6749#section-2.3.1
[2]: https://tools.ietf.org/html/rfc2617#section-2