Handle redirect_uri according to OAuth 2 spec
Passing redirect_uri to authorization server at grant time is optional
(though encouraged and very commonly enforced by providers). However, if
it is passed (as it often is) it is REQUIRED to be passed when using the
grant code to get an access token.

This change allows a user to optionally provide a redirect_uri when
configuring the client to have it passed as a parameter in both the
redirect to the authorization server and the later token exchange.