Remove platform-specific gem version requirements
This needed to be changed so we could upgrade Faraday.

The requirements specified in the Gemspec are fine, but they get
overridden by these Ruby/JRuby 1.8 requirements in the Gemfile.
This is unexpected, and it turns out that you can't specify
different version requirements for different platforms in the Gemfile,
so I'm removing them so the Gemspec's requirements are respected.