Remove OAuth2::Response having an OAuth2::Error as an attribute
In addition to causing circular dependency errors, it is inherently
sort of weird that the place this was being added was in files other
than itself.  While it may be useful to access .response from an error
that you've intercepted, accessing an instance of an error from the
response itself seems less useful.  There were also no specs in Response
asserting the behavior.