Add spec for sites with path components
This was raised in issues 245 and 386, the latter specificially regarding wordpress blogs which often define a "site" as including a path component.

This behaviour is already supported by Faraday (the underlying HTTP client) so long as you specify relative paths when constructing the new URLs rather than absolute ones, e.g. "oauth/authorize" will work but "oauth/token" will not.

It might be desirable to handle this automatically for users by stripping leading `/` characters from the authorize and token URL options (which should really be called `*_path`?), but could also be handled through documentation.

This commit adds a spec demonstrating the use of relative paths as per issue 386.