Add the ability to use TLS client auth when requesting access token
When making the request to retrieve an access token, if the API requires
you to use TLS client auth as the authentication mechanism then it will
fail if the client secret is sent. For this auth type, only the client
ID should be sent.

We do not currently allow the developer to customize this; we set the
client ID and secret by default for all access token requests.

This PR adds an extra section to the case statement in
Authenticator#apply to add the desired functionality; when passing in
`tls_client_auth` as the auth mode, the client secret will be omitted
from the request.