Backport extract_access_token option to OAuth2::Client
This pulls in the changes in #518 from the `1-4-stable` and deprecates
the `extract_access_token` option in favor of the
`access_token_class`.

We move the `access_token_class` as a keyword argument to `get_token`
to avoid having to do dynamic argument checking due to the legacy
`extract_access_token` argument.

`access_token_class` is now responsible for implementing two methods:

- `.from_hash`
- `.contains_token?`

Closes #582