Drop support for MAC tokens (#589)
* Drop support for MAC tokens

MAC tokens never worked properly since the secret was never extracted
from the authorization
payload. https://datatracker.ietf.org/doc/html/draft-ietf-oauth-v2-http-mac-05
never made it as a standard, and as far as I can tell, there is
virtually no support on the server side for this either.

Closes #248

Signed-off-by: Stan Hu <stanhu@gmail.com>

* Regenerate Rubocop TODO

* Disable Rubocop rules in Client#get_token

This is a pre-existing condition.